splited math and autodiff
